:green_heart:: Add mutltiple version test
Use local webpack v3 for test
