+++ __tests__/fixtures/src/index.js
