+++ __tests__/fixtures/webpack.config.js
@@ -1,4 +1,3 @@
-const webpack = require("webpack");
 const path = require("path");
 const LicenseInfoWebpackPlugin = require("../../src/index").default;
