+++ __tests__/v3/package.json
@@ -1,0 +1,8 @@
+{
+  "devDependencies": {
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.3",
+    "babel-preset-env": "^1.6.1",
+    "webpack": "^3.0"
+  }
+}
