+++ package.json
@@ -9,9 +9,10 @@
   "scripts": {
     "watch": "babel -w src -d lib",
     "build": "NODE_ENV=production babel src -d lib",
-    "eslint": "eslint src",
     "prettier": "prettier --write '{src,__tests__}/**/*.js'",
-    "test": "jest",
+    "fixtures:v3": "cd __tests__/v3/ && yarn install",
+    "jest": "jest",
+    "test": "npm-run-all fixtures:* jest",
     "version": "npm run build"
   },
   "author": "Takahiro Abe",
@@ -25,18 +26,21 @@
     "license",
     "plugin"
   ],
-  "peerDependencies": {
-    "webpack": "^3.0"
+  "jest": {
+    "testMatch": [
+      "**/?(*.)(spec|test).js?(x)",
+      "**/__tests__/**/?(*.)(spec|test).js?(x)"
+    ]
   },
+  "peerDependencies": {},
   "devDependencies": {
     "babel-cli": "^6.26.0",
     "babel-core": "^6.0.0",
     "babel-jest": "^21.2.0",
-    "babel-loader": "^7.1.3",
     "babel-preset-env": "^1.1.8",
     "jest": "^21.2.1",
-    "prettier": "^1.9.2",
-    "webpack": "^3.0"
+    "npm-run-all": "^4.1.2",
+    "prettier": "^1.9.2"
   },
   "dependencies": {
     "glob": "^7.1.1",
