+++ src/index.test.js
@@ -201,56 +201,4 @@ Permission is hereby granted, free of charge, to any person obtaining a copy</pr
 `;
     expect(results[0]).toBe(expected);
   });
-
-  test("license-info-webpack-plugin should compile", done => {
-    const webpack = require("webpack");
-    const webpackConfig = require("../test/fixtures/webpack.config.js");
-    const expectedBanner = `/*!
- * vendor@1.0.0 (MIT)
- *   author: yami_beta
- *   url: https://github.com/yami-beta/vendor.git
- *   maintainers:
- *     m1 <m1@example.com> (m1.example.com)
- *     m2 <m2@example.com> (m2.example.com)
- *   contributors:
- *     c1 <c1@example.com> (c1.example.com)
- *     c2 <c2@example.com> (c2.example.com)
- *
- *   MIT License
- *
- *   Copyright (c) 2017 yami_beta
- *
- *   Permission is hereby granted, free of charge, to any person obtaining a copy
- *   of this software and associated documentation files (the "Software"), to deal
- *   in the Software without restriction, including without limitation the rights
- *   to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- *   copies of the Software, and to permit persons to whom the Software is
- *   furnished to do so, subject to the following conditions:
- *
- *   The above copyright notice and this permission notice shall be included in all
- *   copies or substantial portions of the Software.
- *
- *   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- *   IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- *   FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- *   AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- *   LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- *   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- *   SOFTWARE.
- *`;
-
-    webpack(webpackConfig, (err, stats) => {
-      expect(err).toBeFalsy();
-      expect(JSON.stringify(stats.compilation.errors)).toBe("[]");
-      const compiled = stats.compilation.assets["index.js"]
-        .source()
-        .split("\n");
-      const expected = expectedBanner.split("\n");
-      for (let i = 0; i < expected.length; i++) {
-        expect(compiled[i]).toBe(expected[i]);
-      }
-
-      done();
-    });
-  });
 });
